notifications: skip reply pushes for echo-suppressed lifecycle prompts
